rename sharpened to regular
